+++ SomnusFrontEnd/src/app/services/controllers/comment-controller.service.ts
@@ -38,4 +38,9 @@ export class CommentService extends TemplateControllerService<UserComment> {
         let url = `${this.getApiUrlObject()}/add-like/${commentId}`;
         return this.http.get(url);
     }
+
+    public removeLike(commentId: string): Observable<any> {
+        let url = `${this.getApiUrlObject()}/remove-like/${commentId}`;
+        return this.http.get(url);
+    }
 }